refactor Divide form into components, add styling to JobCard, and implement validation
