+++ package.json
@@ -57,6 +57,7 @@
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "@types/react-hooks-helper": "^1.6.4",
+    "@types/react-select": "^5.0.1",
     "@types/styled-components": "^5.1.25",
     "@types/uuid": "^8.3.4",
     "autoprefixer": "^10.4.7",
