+++ src/hooks/useApi.ts
@@ -47,20 +47,23 @@ export const useApi = () => ({
         return res.data;
     },
 
-    createJob: async (title: string, description: string, prerequisites: string, benefits: string, type: string, type_contract: string, salary: number, modality: string, headquarters: string, contract_time: string | boolean, affirmative: string | boolean , affirmative_type: string | undefined, company_id: string | undefined) => {
+    createJob: async (title: string, description: string, prerequisites: string, benefits: string, type: string, typeContract: string, salaryMin: number, salaryMax: number, modality: string, federalUnit: string, city: string, indefinideContract: string | boolean, contractType: string | undefined, affirmative: string | boolean, affirmativeType: string | undefined, company_id: string | undefined) => {
         const res: any = await api.post('/job', {
             title, 
             description, 
             prerequisites, 
             benefits, 
             type, 
-            type_contract, 
-            salary, 
+            typeContract, 
+            salaryMin,
+            salaryMax,
             modality, 
-            headquarters, 
-            contract_time, 
+            federalUnit,
+            city,
+            indefinideContract,
+            contractType,
             affirmative, 
-            affirmative_type,
+            affirmativeType,
             company_id
         })
         return res.data;
